Split ITs
